nits: type fix

+++ csrc/cuda_ops/indices_copy_pad.cu
@@ -44,7 +44,7 @@ void fused_copy_and_pad(
     torch::Tensor in_m_indices,
     torch::Tensor out_hiddens,
     torch::Tensor out_m_indices,
-    int bs, int bucket_bs
+    int64_t bs, int64_t bucket_bs
 ) {
     int hidden_size = in_hiddens.size(1);
     
@@ -61,7 +61,7 @@ void fused_copy_and_pad(
         in_m_indices.data_ptr<int32_t>(),
         reinterpret_cast<uint16_t*>(out_hiddens.data_ptr()),
         out_m_indices.data_ptr<int32_t>(),
-        bs, bucket_bs, hidden_size
+        (int)bs, (int)bucket_bs, hidden_size
     );
 }
