+++ csrc/include/indices_copy_pad.cuh
@@ -7,5 +7,5 @@ void fused_copy_and_pad(
     torch::Tensor in_m_indices,
     torch::Tensor out_hiddens,
     torch::Tensor out_m_indices,
-    int bs, int bucket_bs);
+    int64_t bs, int64_t bucket_bs);
